Updated MaritimeCloudUnit to return a long mmsi instead of integer, fixed relevant issues, updated pom file to work with pluginmanagement for M2Eclipse issues, removed unused msdl, updated plugin repo for msdl generation

+++ epd-common/src/main/java/dk/dma/epd/common/prototype/notification/ChatNotification.java
@@ -72,7 +72,7 @@ public class ChatNotification extends Notification<ChatServiceData, MaritimeId> 
         
         // Try to determine the position
         AisHandlerCommon aisHandler = EPD.getInstance().getAisHandler();
-        Integer mmsi = MaritimeCloudUtils.toMmsi(chatData.getId());
+        Long mmsi = MaritimeCloudUtils.toMmsi(chatData.getId());
         if (mmsi != null && 
             aisHandler.getVesselTarget(mmsi.longValue()) != null &&
             aisHandler.getVesselTarget(mmsi.longValue()).getPositionData() != null) {
